+++ Content.Server/Commands/Observer/Ghost.cs
@@ -27,71 +27,11 @@ namespace Content.Server.Commands.Observer
         {
             if (player == null)
             {
-                shell.SendText(player, "Nah");
+                shell?.SendText((IPlayerSession?) null, "Nah");
                 return;
             }
 
-            var mind = player.ContentData()?.Mind;
-
-            if (mind == null)
-            {
-                shell.SendText(player, "You can't ghost here!");
-                return;
-            }
-
-            var playerEntity = player.AttachedEntity;
-
-            if (playerEntity != null && playerEntity.HasComponent<GhostComponent>())
-                return;
-
-            if (mind.VisitingEntity != null)
-            {
-                mind.UnVisit();
-                mind.VisitingEntity.Delete();
-            }
-
-            var position = playerEntity?.Transform.Coordinates ?? IoCManager.Resolve<IGameTicker>().GetObserverSpawnPoint();
-            var canReturn = false;
-
-            if (playerEntity != null && CanReturn && playerEntity.TryGetComponent(out IMobStateComponent? mobState))
-            {
-                if (mobState.IsDead())
-                {
-                    canReturn = true;
-                }
-                else if (mobState.IsCritical())
-                {
-                    canReturn = true;
-
-                    if (playerEntity.TryGetComponent(out IDamageableComponent? damageable))
-                    {
-                        //todo: what if they dont breathe lol
-                        damageable.ChangeDamage(DamageType.Asphyxiation, 100, true);
-                    }
-                }
-                else
-                {
-                    canReturn = false;
-                }
-            }
-
-            var entityManager = IoCManager.Resolve<IEntityManager>();
-            var ghost = entityManager.SpawnEntity("MobObserver", position);
-            ghost.Name = mind.CharacterName;
-
-            var ghostComponent = ghost.GetComponent<GhostComponent>();
-            ghostComponent.CanReturnToBody = canReturn;
-
-            if (playerEntity != null &&
-                playerEntity.TryGetComponent(out ServerOverlayEffectsComponent? overlayComponent))
-            {
-                overlayComponent.RemoveOverlay(SharedOverlayID.CircleMaskOverlay);
-            }
-
-            if (canReturn)
-                mind.Visit(ghost);
-            else
-                mind.TransferTo(ghost);
+            IoCManager.Resolve<IGameTicker>().OnGhostAttempt(shell, player, CanReturn);
         }
     }
 }
