+++ Content.Server/GameTicking/GamePreset.cs
@@ -1,8 +1,20 @@
+#nullable enable annotations
 ﻿using System.Collections.Generic;
 using Content.Shared.Preferences;
-using Robust.Server.Interfaces.Player;
+using Content.Server.Administration;
+using Content.Server.GameObjects.Components.Mobs;
+using Content.Server.GameObjects.Components.Observer;
+using Content.Server.Interfaces.GameTicking;
+using Content.Server.Players;
+using Content.Shared.Damage;
+using Content.Shared.GameObjects.Components.Damage;
+using Content.Shared.GameObjects.Components.Mobs;
+using Content.Shared.GameObjects.Components.Mobs.State;
 using Robust.Shared.Network;
 using Robust.Shared.Interfaces.GameObjects;
+using Robust.Server.Interfaces.Console;
+﻿using Robust.Server.Interfaces.Player;
+using Robust.Shared.IoC;
 
 namespace Content.Server.GameTicking
 {
@@ -23,6 +35,75 @@ namespace Content.Server.GameTicking
         /// Called when a player is spawned in (this includes, but is not limited to, before Start)
         /// </summary>
         public virtual void OnSpawnPlayerCompleted(IPlayerSession session, IEntity mob, bool lateJoin) { }
+
+        /// <summary>
+        /// Called when a player attempts to ghost.
+        /// </summary>
+        public virtual void OnGhostAttempt(IConsoleShell? shell, IPlayerSession player, bool canReturnGlobal)
+        {
+            var mind = player.ContentData().Mind;
+            if (mind == null)
+            {
+                shell?.SendText(player, "You can't ghost here!");
+                return;
+            }
+
+            var name = player.AttachedEntity?.Name ?? player.Name;
+
+            var playerEntity = player.AttachedEntity;
+
+            if (playerEntity != null && playerEntity.HasComponent<GhostComponent>())
+                return;
+
+            if (mind.VisitingEntity != null)
+            {
+                mind.UnVisit();
+                mind.VisitingEntity.Delete();
+            }
+
+            var position = playerEntity?.Transform.Coordinates ?? IoCManager.Resolve<IGameTicker>().GetObserverSpawnPoint();
+            var canReturn = false;
+
+            if (playerEntity != null && canReturnGlobal && playerEntity.TryGetComponent(out IMobStateComponent? mobState))
+            {
+                if (mobState.IsDead())
+                {
+                    canReturn = true;
+                }
+                else if (mobState.IsCritical())
+                {
+                    canReturn = true;
+
+                    if (playerEntity.TryGetComponent(out IDamageableComponent? damageable))
+                    {
+                        //todo: what if they dont breathe lol
+                        damageable.ChangeDamage(DamageType.Asphyxiation, 100, true);
+                    }
+                }
+                else
+                {
+                    canReturn = false;
+                }
+            }
+
+            var entityManager = IoCManager.Resolve<IEntityManager>();
+            var ghost = entityManager.SpawnEntity("MobObserver", position);
+            ghost.Name = mind.CharacterName;
+
+            var ghostComponent = ghost.GetComponent<GhostComponent>();
+            ghostComponent.CanReturnToBody = canReturn;
+
+            if (playerEntity != null &&
+                playerEntity.TryGetComponent(out ServerOverlayEffectsComponent? overlayComponent))
+            {
+                overlayComponent.RemoveOverlay(SharedOverlayID.CircleMaskOverlay);
+            }
+
+            if (canReturn)
+                mind.Visit(ghost);
+            else
+                mind.TransferTo(ghost);
+        }
 
         public virtual string GetRoundEndDescription() => "";
     }
